+++ contracts/protocols/cdps/sky/interfaces/dss/mip21/RwaInputConduitAbstract.sol
@@ -35,9 +35,9 @@ interface RwaInputConduit3Abstract is RwaInputConduitBaseAbstract {
     function gem() external view returns (address);
     function quitTo() external view returns (address);
     function file(bytes32, address) external;
-    function push(uint) external;
+    function push(uint256) external;
     function quit() external;
-    function quit(uint) external;
+    function quit(uint256) external;
     function yank(address, address, uint256) external;
     function expectedDaiWad(uint256) external view returns (uint256);
     function requiredGemAmt(uint256) external view returns (uint256);
